Fix: United the context and functionality code all over the website
