+++ app/components/ProjectInfo.js
@@ -1,8 +1,6 @@
 import React, { useEffect, useState } from "react";
 import user_img from "../../public/assets/images/user_img.jpg";
-import project_img from "../../public/assets/images/project.jpg";
-import frame_hand from "../../public/assets/images/frame_hand.png";
-import handImag from "../../public/assets/images/hand.png";
+
 import Image from "next/image";
 import DonationForm from "./forms/DonationForm";
 import { useAuth } from "@/context/AuthContext";
@@ -20,6 +18,7 @@ export const ProjectInfo = ({
   const [openDonationForm, setOpenDonationForm] = useState(false);
   const [daysLeft, setDaysLeft] = useState(null);
   const { user } = useAuth();
+  
   useEffect(() => {
     const endDate = new Date(left);
     const today = new Date();
@@ -36,39 +35,7 @@ export const ProjectInfo = ({
       : setOpenDonationForm(false);
   };
 
-  const styles = {
-    project_size: {
-      minHeight: "600px",
-      borderRadius: "10px",
-    },
 
-    profile_size: {
-      height: "40px",
-      width: "40px",
-      borderRadius: "10px",
-    },
-    project_class: `h-full w-auto object-cover `,
-    page: `grid md:grid-cols-1 lg:grid-cols-4 xl:grid-cols-5 mx-5 my-2`,
-    left: `flex ml-10 mt-12 col-span-2  rounded-3xl justify-center items-center`,
-    right: `xl:col-span-3 lg:col-span-2 mx-8 mt-12`,
-    short_profile: `flex mb-10`,
-    prof_title: `text-4xl mb-10 text-left lg:me-24 xl:me-48`,
-    prof_name: `ml-5 mt-2 font-bold text-orange-800`,
-    inproject: `grid grid-cols-2 border-y-2 border-black`,
-    intoleft: `col-span-1 my-8`,
-    intoright: `col-span-1 my-8 border-s-2 border-black`,
-    about_title: `text-lg text-left`,
-    about: `text-sm text-left mt-5 2xl:me-40 xl:me-20 me-10 text-gray-500`,
-    ul: `xl:flex justify-around text-start mx-3`,
-    raised_li: `py-2 px-4 text-sm`,
-    goal_li: `bg-orange-500 py-2 px-4 rounded-md text-sm`,
-    numbers: `text-2xl`,
-    percent: `bg-gray-200 w-5/6 h-3 rounded-md left-10 mt-10 mx-auto`,
-    onpercent: `bg-orange-500 w-5/6 h-3 rounded-md left-10`,
-    left_day: `flex mt-8 justify-center text-sm`,
-    button: `bg-black hover:bg-orange-500 text-white font-bold py-2 px-4 rounded-md border border-black mt-14 w-80 text-center ms-0 md:w-100`,
-    button_pos: `flex justify-center md:justify-start`,
-  };
 
   return (
     <div className={styles.page}>
@@ -156,6 +123,37 @@ export const ProjectInfo = ({
   );
 };
 
-//<img src={frame_hand} className='z-10 ' style={styles.frame_size} ></img>
-// if objects get wanted..
-// frame_size: { minHeight: "400px", height: "820px", objectfit: "cover"}
+ 
+const styles = {
+  project_size: {
+    minHeight: "600px",
+    borderRadius: "10px",
+  },
+
+  profile_size: {
+    height: "40px",
+    width: "40px",
+    borderRadius: "10px",
+  },
+  project_class: `h-full w-auto object-cover `,
+  page: `grid md:grid-cols-1 lg:grid-cols-4 xl:grid-cols-5 mx-5 my-2`,
+  left: `flex ml-10 mt-12 col-span-2  rounded-3xl justify-center items-center`,
+  right: `xl:col-span-3 lg:col-span-2 mx-8 mt-12`,
+  short_profile: `flex mb-10`,
+  prof_title: `text-4xl mb-10 text-left lg:me-24 xl:me-48`,
+  prof_name: `ml-5 mt-2 font-bold text-orange-800`,
+  inproject: `grid grid-cols-2 border-y-2 border-black`,
+  intoleft: `col-span-1 my-8`,
+  intoright: `col-span-1 my-8 border-s-2 border-black`,
+  about_title: `text-lg text-left`,
+  about: `text-sm text-left mt-5 2xl:me-40 xl:me-20 me-10 text-gray-500`,
+  ul: `xl:flex justify-around text-start mx-3`,
+  raised_li: `py-2 px-4 text-sm`,
+  goal_li: `bg-orange-500 py-2 px-4 rounded-md text-sm`,
+  numbers: `text-2xl`,
+  percent: `bg-gray-200 w-5/6 h-3 rounded-md left-10 mt-10 mx-auto`,
+  onpercent: `bg-orange-500 w-5/6 h-3 rounded-md left-10`,
+  left_day: `flex mt-8 justify-center text-sm`,
+  button: `bg-black hover:bg-orange-500 text-white font-bold py-2 px-4 rounded-md border border-black mt-14 w-80 text-center ms-0 md:w-100`,
+  button_pos: `flex justify-center md:justify-start`,
+};
