+++ app/components/category/CategoryFiltering.jsx
@@ -1,5 +1,5 @@
 import { Box, Typography } from "@mui/material";
-import Image from "next/image";
+
 import React, { useState } from "react";
 
 const styles = {
