+++ app/context/AuthContext.jsx
@@ -6,13 +6,28 @@ import {
   signInWithEmailAndPassword,
   signOut,
 } from "firebase/auth";
-import { auth } from "../config/firebase";
+import {
+  collection,
+  onSnapshot,
+  query,
+  doc,
+  deleteDoc,
+} from "firebase/firestore";
+import { auth, db } from "../config/firebase";
+
 const AuthContext = createContext();
 
 export const AuthContextProvider = ({ children }) => {
   const [user, setUser] = useState(null);
   const [loading, setLoading] = useState(true);
+  const [usersInfo, setUsersInfo] = useState(null);
+  const [projects, setProjects] = useState(true);
+  const [donations, setDonations] = useState([]);
+  const [currentUser, setCurrentUser] = useState(null);
+  const [usersProjects, setUsersProjects] = useState([]);
 
+
+  //Auth CONTEXT
   useEffect(() => {
     const unsubscribe = onAuthStateChanged(auth, (user) => {
       if (user) {
@@ -45,10 +60,92 @@ export const AuthContextProvider = ({ children }) => {
     setUser(null);
     await signOut(auth);
   };
+  const deleteProject = async (oneProjectInfo) => {
+    await deleteDoc(doc(db, "projects", oneProjectInfo.id));
+    console.log("item deleted");
+  };
+  //FUND CONTEXT
+  useEffect(() => {
+    const q = query(collection(db, "users"));
+    const unsubscribe = onSnapshot(q, (QuerySnapshot) => {
+      let usersArr = [];
+      QuerySnapshot.forEach((doc) => {
+        usersArr.push({ ...doc.data(), id: doc.id });
+      });
+      setUsersInfo(usersArr);
+    });
 
+    return () => unsubscribe();
+  }, []);
+
+  useEffect(() => {
+    const q = query(collection(db, "projects"));
+    const unsubscribe = onSnapshot(q, (QuerySnapshot) => {
+      let projectsArr = [];
+      QuerySnapshot.forEach((doc) => {
+        projectsArr.push({ ...doc.data(), id: doc.id });
+      });
+      setProjects(projectsArr);
+    });
+    return () => unsubscribe();
+  }, []);
+
+  useEffect(() => {
+    const q = query(collection(db, "donations"));
+    const unsubscribe = onSnapshot(q, (QuerySnapshot) => {
+      let donationsArr = [];
+      QuerySnapshot.forEach((doc) => {
+        donationsArr.push({ ...doc.data(), id: doc.id });
+      });
+      setDonations(donationsArr);
+    });
+    return () => unsubscribe();
+  }, []);
+
+  //current user state
+  useEffect(() => {
+    const fetchData = async () => {
+      if (user && usersInfo) {
+        // Find the current user's data and set it in the currentUser state
+        const currentUserData = usersInfo.find((x) => x.id === user.uid);
+        setCurrentUser(currentUserData);
+      }
+      setLoading(false);
+    };
+
+    fetchData();
+  }, [currentUser, user, usersInfo]);
+
+  // Single User Projects
+  useEffect(() => {
+    if (user && projects) {
+      const projectArray = Object.values(projects);
+      const projectWithUser = projectArray.filter(
+        (project) => project.creator.userId === user.uid
+      );
+      setUsersProjects(projectWithUser);
+    }
+  }, [projects, user]);
+
+ 
+  
   return (
-    <AuthContext.Provider value={{ user, login, signup, logout, loading }}>
-      {loading ? null : children}
+    <AuthContext.Provider
+      value={{
+        user,
+        login,
+        signup,
+        logout,
+        loading,
+        usersInfo,
+        projects,
+        donations,
+        currentUser,
+        deleteProject,
+        usersProjects,
+      }}
+    >
+      {loading ? null || undefined : children}
     </AuthContext.Provider>
   );
 };
