--- app/context/FundContext.jsx
@@ -1,55 +1,0 @@
-"use client";
-import { createContext, useEffect, useState } from "react";
-import { collection, onSnapshot, query } from "firebase/firestore";
-import { db } from "../config/firebase";
-
-const FundContext = createContext();
-
-const FundProvider = ({ children }) => {
-  const [usersInfo, setUsersInfo] = useState(null);
-  const [projects, setProjects] = useState(true);
-  const [donations, setDonations] = useState([]);
-  useEffect(() => {
-    const q = query(collection(db, "users"));
-    const unsubscribe = onSnapshot(q, (QuerySnapshot) => {
-      let usersArr = [];
-      QuerySnapshot.forEach((doc) => {
-        usersArr.push({ ...doc.data(), id: doc.id });
-      });
-      setUsersInfo(usersArr);
-    });
-    return () => unsubscribe();
-  }, []);
-
-  useEffect(() => {
-    const q = query(collection(db, "projects"));
-    const unsubscribe = onSnapshot(q, (QuerySnapshot) => {
-      let projectsArr = [];
-      QuerySnapshot.forEach((doc) => {
-        projectsArr.push({ ...doc.data(), id: doc.id });
-      });
-      setProjects(projectsArr);
-    });
-    return () => unsubscribe();
-  }, []);
-
-  useEffect(() => {
-    const q = query(collection(db, "donations"));
-    const unsubscribe = onSnapshot(q, (QuerySnapshot) => {
-      let donationsArr = [];
-      QuerySnapshot.forEach((doc) => {
-        donationsArr.push({ ...doc.data(), id: doc.id });
-      });
-      setDonations(donationsArr);
-    });
-    return () => unsubscribe();
-  }, []);
-
-  return (
-    <FundContext.Provider value={{ usersInfo, projects, donations }}>
-      {children}
-    </FundContext.Provider>
-  );
-};
-
-export { FundContext, FundProvider };
